backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
python-docx
spacy
requests
beautifulsoup4
selenium
fake-useragent
sentence-transformers
scikit-learn
numpy
pandas
python-dotenv
aiofiles
Pillow
paddleocr
paddlepaddle
webdriver-manager
python-jose
passlib
bcrypt

# Additional requirements for job scraping
urllib3>=1.26.0
lxml>=4.6.0
pydantic>=2.0.0
python-dateutil>=2.8.0
pytz>=2023.3

# Optional but recommended
httpx>=0.24.0
cloudscraper>=1.2.60
aiohttp>=3.8.0

geoip2>=4.0.0